use azure-scheduled-events app for draining nodes during an upgrade
